Uses a HTTP GET to request the flight prices
Works as of now for adult prices. Result parser choked when there were
senior tickets. It was able tro get the prices for tickets to
destination, but not on return. The return did have one set of tickets
that were sold old. Glancing at that part of the code the senior ticket
column did not appear to be scanned anyways.

Previously used argment values for time and date are incompatible with
the new form so some helper functions had to be added to convert their
formats.